refactor(auth): extract sign out
